# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ModelForge Public API",
    "version": "1.0.0",
    "description": "The open deployment intelligence layer between Hugging Face models and production AI compute. Powered by OpenComputeBench reproducible inference benchmarks.",
    "contact": {
      "name": "ModelForge Team",
      "url": "https://modelforge.dev",
      "email": "api@modelforge.dev"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://modelforge.dev/api",
      "description": "Production Cloud API"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/v1/models/{modelId}/passport": {
      "get": {
        "summary": "Get Model Compute Passport",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "URL-encoded Hugging Face model repository ID"
          },
          {
            "name": "revision",
            "in": "query",
            "schema": { "type": "string", "default": "main" },
            "description": "Exact commit SHA or branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Compute Passport record with confidence scoring",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComputePassport" }
              }
            }
          },
          "404": {
            "description": "Model or passport not found"
          }
        }
      }
    },
...
    "/v1/benchmarks": {
      "get": {
        "summary": "Query OpenComputeBench Records",
        "parameters": [
          { "name": "model", "in": "query", "schema": { "type": "string" } },
          { "name": "hardware", "in": "query", "schema": { "type": "string" } },
          { "name": "runtime", "in": "query", "schema": { "type": "string" } },
          {
            "name": "precision",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "verifiedOnly",
            "in": "query",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of benchmark records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BenchmarkRecord" }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/benchmark-submissions": {
      "post": {
        "summary": "Submit Benchmark Observation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BenchmarkRecord" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Benchmark accepted and verified"
          },
          "400": {
            "description": "Validation or cryptographic hash mismatch"
          }
        }
      }
    },
    "/v1/failures": {
      "get": {
        "summary": "Query Deployment Failure Corpus",
        "parameters": [
          { "name": "model", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "runtime", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Normalized failure records with mitigations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_count": { "type": "integer" },
                    "failures": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/FailureRecord" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComputePassport": {
        "type": "object",
        "properties": {
          "model_id": { "type": "string" },
          "revision": { "type": "string" },
          "confidence_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "confidence_explanation": { "type": "string" },
          "compatibility": { "type": "object" },
          "memory_profile": { "type": "object" }
        }
      },
...
      "BenchmarkRecord": {
        "type": "object",
        "required": [
          "benchmark_id",
          "schema_version",
          "model",
          "hardware",
          "runtime",
          "metrics",
          "provenance"
        ],
        "properties": {
          "benchmark_id": { "type": "string", "format": "uuid" },
          "schema_version": { "type": "string" },
          "golden": { "type": "boolean" },
          "synthetic_fixture": { "type": "boolean" },
          "model": { "type": "object" },
          "hardware": { "type": "object" },
          "runtime": { "type": "object" },
          "metrics": { "type": "object" },
          "provenance": { "type": "object" }
        }
      },
      "FailureRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "model_repository": { "type": "string" },
          "model_revision": { "type": "string" },
          "runtime": { "type": "string" },
          "accelerator": { "type": "string" },
          "failure_category": { "type": "string" },
          "normalized_reason": { "type": "string" },
          "mitigation": { "type": "string" }
        }
      }
    }
  }
}